<release version="1.1.29" versioncode="38" date="30/04/2017">         <change>Possibility to filter the graphs by type to display in statistics activity.</change>         <change>Add openvpn - autoshutdown - fail2ban - transmissionbt - in logs activity.</change>     </release>     <release version="1.1.28" versioncode="37" date="27/04/2017">         <change>Add Fail2ban activity.</change>         <change>Code improvement.</change>     </release>

+++ src/main/java/DAL/HostsTable.java
@@ -3,6 +3,7 @@ package DAL;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
 
+
 /**
  * Created by thiba on 11/10/2016.
  */
@@ -42,11 +43,15 @@ public class HostsTable {
 
     public static void onUpgrade(SQLiteDatabase database, int oldVersion,
                                  int newVersion) {
-        Log.w(HostsTable.class.getName(), "Upgrading database from version "
-                + oldVersion + " to " + newVersion
-                + ", which will destroy all old data");
-        database.execSQL("DROP TABLE IF EXISTS " + TABLE_HOST);
-        onCreate(database);
+
+
+        if(false) {
+            Log.w(HostsTable.class.getName(), "Upgrading database from version "
+                    + oldVersion + " to " + newVersion
+                    + ", which will destroy all old data");
+            database.execSQL("DROP TABLE IF EXISTS " + TABLE_HOST);
+            onCreate(database);
+        }
     }
 
 }
